content: cut down header
Rather than continuing the discussion at
https://github.com/cryptomator/new-encryption-format/discussions ,
I would like to try make things more concrete via PRs like this one.

So please take this as an RFC i.e. a base for discussion.

For the fields I deleted:

* file body offset

This seems only be useful if the file is split into multiple parts.
I don't think we want to do this, as it makes many operations non-
atomic

* file key information

This seems to be covered already in what I called "FILE ID"

* file body format (cipher mode, block size, ...)

I don't think we want this configurable at a per-file level.

This should be fixed, or, if we need it configurable, it should
be in the config file.